IGNITE-24499 Extract Primary Replica negotiation logic from ReplicaImpl do dedicated class
